// src/enums/BaseObjects.ts
export enum BaseObjects {
  ARRAY = 'ArrayObject',
  BOOLEAN = 'BooleanObject',
  DICTIONARY = 'DictionaryObject',
  NAME = 'NameObject',
  NULL = 'NullObject',
  NUMERIC = 'NumericObject',
  STREAM = 'StreamObject',
  STRING = 'StringObject',
  INTEGER = 'IntegerObject',
}

export default BaseObjects;